refactor instruction
