+++ mcp-sqlServer/local.settings.json
@@ -6,7 +6,6 @@
         "SQL_SERVER": "azuresdkbi-server.database.windows.net",
         "SQL_DATABASE": "azuresdkbi",
         "AZURE_SUBSCRIPTION_ID": "a18897a6-7e44-457d-9260-f2854c0aca42",
-        "AZURE_RESOURCE_GROUP": "sdk-mgmt-bi-data",
-        "MCP_SERVER": "AI"
+        "AZURE_RESOURCE_GROUP": "sdk-mgmt-bi-data"
     }
 }
