--- mcp-sqlServer/reference/mappings/empty
@@ -1,1 +1,0 @@
-for further data set multi alias
