+++ mcp-sqlServer/src/schemas/KQL_Schema.json
