+++ mcp-sqlServer/src/schemas/SQL_Schema.json
