+++ mcp-sqlServer/src/services/kusto_mcp_tools.py
@@ -10,14 +10,14 @@ class KQLGeneratorMCP:
     """KQL Generator MCP tool for template-based query generation"""
     
     def __init__(self):
-        self.sample_kql_path = os.path.join(
-            os.path.dirname(os.path.dirname(os.path.dirname(__file__))), 
-            "reference", "samples", "sample.kql"
+        self.kql_template_path = os.path.join(
+            os.path.dirname(os.path.dirname(__file__)), 
+            "templates", "queries", "azure_sdk_usage_query.kql"
         )
     
     async def generate_kql_from_template(self, user_question: str) -> Dict[str, Any]:
         """
-        Based on sample.kql template and user question, return structured data for AI processing
+        Based on Azure SDK usage KQL template and user question, return structured data for AI processing
         
         Args:
             user_question: User's query requirement
@@ -26,14 +26,12 @@ class KQLGeneratorMCP:
             Structured data containing template and user question for AI processing
         """
         try:
-            # Read sample.kql template
-            sample_kql = self._read_sample_kql()
+            kql_template = self._read_kql_template()
             
-            # Return structured data for AI
             return {
                 "success": True,
                 "user_question": user_question,
-                "sample_kql_template": sample_kql,
+                "kql_template": kql_template,
                 "instructions": {
                     "task": "Generate a new KQL query based on the user question and sample template",
                     "requirements": [
@@ -68,15 +66,15 @@ class KQLGeneratorMCP:
                 "success": False,
                 "error": f"Error processing request: {str(e)}",
                 "user_question": user_question,
-                "suggestion": "Please check if sample.kql file exists and is readable"
+                "suggestion": "Please check if azure_sdk_usage_query.kql template file exists and is readable"
             }
     
-    def _read_sample_kql(self) -> str:
-        """Read content of sample.kql file"""
+    def _read_kql_template(self) -> str:
+        """Read content of azure_sdk_usage_query.kql template file"""
         try:
-            with open(self.sample_kql_path, 'r', encoding='utf-8') as f:
+            with open(self.kql_template_path, 'r', encoding='utf-8') as f:
                 return f.read()
         except FileNotFoundError:
-            raise Exception(f"Sample KQL file not found: {self.sample_kql_path}")
+            raise Exception(f"KQL template file not found: {self.kql_template_path}")
         except Exception as e:
-            raise Exception(f"Error reading sample KQL file: {str(e)}")
+            raise Exception(f"Error reading KQL template file: {str(e)}")
