+++ mcp-sqlServer/src/templates/prompts/QueryTypeDecision.md
