+++ mcp-sqlServer/src/templates/queries/azure_sdk_usage_query.kql
